chore: rename residual_buffer to residual
