+++ src/cala/streaming/nodes/init/odl/__init__.py
@@ -1,7 +1,7 @@
 from .component_stats import ComponentStatsInitializer
 from .overlaps import OverlapsInitializer
 from .pixel_stats import PixelStatsInitializer
-from .residual_buffer import ResidualInitializer
+from .residuals import ResidualInitializer
 
 __all__ = [
     "PixelStatsInitializer",
